Authentication added

+++ E-shop-project/backend/routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\CategoriesController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\ProductsController;
@@ -30,23 +31,36 @@ Route::group(['prefix' => 'products'], function() {
     Route::get('/', [ProductsController::class, 'index']);
     Route::get('/s/{keyword}', [ProductsController::class, 'search']);
     Route::get('/{field}/{order}', [ProductsController::class, 'order']);
-    Route::get('/{id}', [ProductsController::class, 'singleProduct'])->where('id', '[0-9]+');
-    Route::post('/', [ProductsController::class, 'create']);
-    Route::put('/{id}', [ProductsController::class, 'update'])->where('id', '[0-9]+');
-    Route::delete('/{id}', [ProductsController::class, 'delete'])->where('id', '[0-9]+');
+
+    // Admino routai
+    Route::middleware('auth:sanctum')->get('/{id}', [ProductsController::class, 'singleProduct'])->where('id', '[0-9]+');
+    Route::middleware('auth:sanctum')->post('/', [ProductsController::class, 'create']);
+    Route::middleware('auth:sanctum')->put('/{id}', [ProductsController::class, 'update'])->where('id', '[0-9]+');
+    Route::middleware('auth:sanctum')->delete('/{id}', [ProductsController::class, 'delete'])->where('id', '[0-9]+');
 });
 
 Route::group(['prefix' => 'categories'], function() {
     Route::get('/', [CategoriesController::class, 'index']);
-    Route::get('/{id}', [CategoriesController::class, 'singleCategory'])->where('id', '[0-9]+');
     Route::get('/products/{id}', [ProductsController::class, 'categoryProducts'])->where('id', '[0-9]+');
-    Route::post('/', [CategoriesController::class, 'create']);
-    Route::delete('/{id}', [CategoriesController::class, 'delete'])->where('id', '[0-9]+');
-    Route::put('/{id}', [CategoriesController::class, 'update'])->where('id', '[0-9]+');
+
+    // Admino routai
+    Route::middleware('auth:sanctum')->post('/', [CategoriesController::class, 'create']);
+    Route::middleware('auth:sanctum')->delete('/{id}', [CategoriesController::class, 'delete'])->where('id', '[0-9]+');
+    Route::middleware('auth:sanctum')->put('/{id}', [CategoriesController::class, 'update'])->where('id', '[0-9]+');
+    Route::middleware('auth:sanctum')->get('/{id}', [CategoriesController::class, 'singleCategory'])->where('id', '[0-9]+');
 });
 
 Route::group(['prefix' => 'orders'], function() {
-    Route::get('/', [OrderController::class, 'index']);
     Route::post('/', [OrderController::class, 'create']);
-    Route::put('/{id}', [OrderController::class, 'update'])->where('id', '[0-9]+');
+
+    // Admino routai
+    Route::middleware('auth:sanctum')->get('/', [OrderController::class, 'index']);
+    Route::middleware('auth:sanctum')->put('/{id}', [OrderController::class, 'update'])->where('id', '[0-9]+');
 });
+
+Route::post('/login', [AuthController::class, 'login']);
+Route::post('/register', [AuthController::class, 'register']);
+
+//Jei insomnia ar postman meta logout errora, nustatyti headeri name: Accept, value: application/json
+Route::middleware('auth:sanctum')->get('/logout', [AuthController::class, 'logout']);
+Route::middleware('auth:sanctum')->get('/check', [AuthController::class, 'index']);
